+++ blockchain-eos/App/AppDelegate.swift
@@ -15,29 +15,21 @@ import OHHTTPStubs
 @UIApplicationMain
 class AppDelegate: UIResponder, UIApplicationDelegate {
     
+    var window: UIWindow?
+    
+    private var appCoordinator: AppCoordinator?
+
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
         // Override point for customization after application launch.
         if CommandLine.arguments.contains("--uitesting") {
             self.stubNetworking()
         }
         
+        start()
+        
         setAppColors()
         
         return true
-    }
-
-    // MARK: UISceneSession Lifecycle
-
-    func application(_ application: UIApplication, configurationForConnecting connectingSceneSession: UISceneSession, options: UIScene.ConnectionOptions) -> UISceneConfiguration {
-        // Called when a new scene session is being created.
-        // Use this method to select a configuration to create the new scene with.
-        return UISceneConfiguration(name: "Default Configuration", sessionRole: connectingSceneSession.role)
-    }
-
-    func application(_ application: UIApplication, didDiscardSceneSessions sceneSessions: Set<UISceneSession>) {
-        // Called when the user discards a scene session.
-        // If any sessions were discarded while the application was not running, this will be called shortly after application:didFinishLaunchingWithOptions.
-        // Use this method to release any resources that were specific to the discarded scenes, as they will not return.
     }
     
     private func setAppColors() {
@@ -46,6 +38,18 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         UITextView.appearance().textColor =  Constants.UI.AppTextColor
     }
 
+    private func start() {
+        
+        let window = UIWindow(frame: UIScreen.main.bounds)
+        self.window = window
+        
+        let navController = UINavigationController()
+        let apiClient = EosApiClient(endpoint: Constants.Networking.EosNodeEndpoint)
+        
+        self.appCoordinator = AppCoordinator(window: window, presenter: navController, apiClient: apiClient)
+        
+        self.appCoordinator?.start()
+    }
     #if DEBUG
     private func stubNetworking() {
